schnapps: update to 2.2.1 and refactor
This refactors this package and cleans it overall.

Hacky generation of updater scripts was removed and instead two scripts
were introduced. This is minimal overhead and is cleaner.

Dependencies were expanded.
schnapps reads uci and uses uci command so it should depend on it.
It also uses crond so dependencies should be there as well.
And lastly it has hook calling cert-backup so dependency should there as
well.