cert-backup: clean and refactor and fix for Turris 1.x
Lede changed to OpenWrt and with that they changed LEDE_DEVICE_PRODUCT
to OPENWRT_DEVICE_PRODUCT. This breaks cert-backup on Turris 1.x. It is
more secure to just decide what backed will be used in package it self
and not in runtime of script. This does that. Mounting of cert-backup
store is split to separate backend script and there are two variants to
deploy to different routers.

This commit also removes licenses. Whole repository is licensed. In
reality we do not have common licensing defined and so it is unsure if
we should have every file licensed or if it is enough to have license on
package and license on repository that package is generated from. I am
choosing the second one because it removed need to update license in
every script.