switch-branch: ensure that reinstall happens on branch switch
This changes original code that just executed updater with reinstall-all
option with more clever code in Lua to detect when we have to reinstall
and when it is not required to reinstall everything.
This fixes situation when updater fails and user later does not run
pkgupdate --reinstall-all. This sets it so even if updater is run
outside of switch-branch script the switch is detected and reinstall
performed.