fix: fix case when updater is run with reinstall-all option
When updater is run with --reinstall-all option it won't reexecute and
that way fix is not executed. What it on the other hand does is that it
runs postupdate hooks. That gives us place to consistently call
update_alternatives.sh.

This fixes only situation when no replan is performed. When replan is
performed then this hook is removed before updater terminates so it is
not called.

This modification of fix package also allows us to add any additional
files to fix package. The approach is simply to define default section
you can call and using default as the one only if no install section was
defined.